+++ src/components/Main.jsx
@@ -3,7 +3,7 @@ import ItemListContainer from "./ItemListContainer";
 import ItemDetail from "./ItemDetail";
 import Home from "./Home";
 import Categories from "./utilities/Categories";
-import Carrito from "./Carrito";
+import Carrito from "./utilities/Carrito";
 
 function Main(){
 
@@ -11,7 +11,7 @@ function Main(){
         <main className="main">
             <Routes>
                 <Route path="/" element={<Home/>}/>
-                <Route path="/product/" element={<ItemListContainer greeting="Cargando Productos..."/>}/>
+                <Route path="/product/" element={<ItemListContainer/>}/>
                 <Route path="/cart" element={<Carrito/>}/>
                 <Route path="/product/:id" element={<ItemDetail/>}/>
                 <Route path="/product/category/:cat" element={<Categories/>}/>
