+++ src/components/utilities/CartWidget.jsx
@@ -1,14 +1,17 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faCartShopping } from '@fortawesome/free-solid-svg-icons'
+import { useState } from 'react';
+import Input from './Input';
 
 
 function CartWidget(){
-
+    const [data, setData] = useState();
+    
     return(
         <>
         <div className='cart'>
             <FontAwesomeIcon icon={faCartShopping} className='cartIcon'/>
-            <p className='cartTotalProduct'>4</p>
+            <Input isCart={true}/>
         </div>
         
         </>
