+++ src/sass/components/_container.scss
@@ -40,9 +40,8 @@
                 border-radius: 10px;
                 transition: .4s;
                 &:active{
-                    transition: .2s;
-                    color: $textPrimary;
-                    background-image: linear-gradient($bgPrincipal,$bgSecundary);
+                    transition: .3s;
+                    background-image: linear-gradient(#88B7B5,#847996);
                 }
             }
         }
