+++ src/sass/components/_productDetails.scss
@@ -37,12 +37,16 @@
             color: $textSecundary;
         }
         .btnCount{
+            @include flex(column, center, center);
             border: 1px solid $bgSecundary;
             padding: 5px;
             border-radius: 10px;
             margin: 0px;
             font-weight: 600;
             font-size: large;
+            button{
+                margin: 5px;
+            }
         }
     }
 }
