--- vite-project/src/components/Footer.jsx
@@ -1,15 +1,0 @@
-import NavBar from "./NavBar";
-
-function Footer(){
-
-    return(
-        <>
-        <div className="footer">
-        <NavBar isFooter={true}></NavBar>
-
-        </div>
-        </>
-    )
-}
-
-export default Footer;
