--- vite-project/src/components/Header.jsx
@@ -1,15 +1,0 @@
-import NavBar from "./NavBar";
-import "../sass/components/_navBar.scss"
-
-function Header(){
-
-    return(
-        <>
-        <header className="header">
-        <NavBar/>
-        </header>
-        </>
-    )
-}
-
-export default Header;
