--- vite-project/src/components/ItemListContainer.jsx
@@ -1,15 +1,0 @@
-import { useEffect, useState } from "react";
-import ItemList from "./utilities/ItemList";
-import Category from "./utilities/Category";
-
-
-function ItemListContainer(){
-
-    return(
-        <div className="container">
-            <ItemList/>
-        </div>
-    )
-}
-
-export default ItemListContainer;
