--- vite-project/src/components/utilities/Category.jsx
@@ -1,25 +1,0 @@
-import { useEffect, useState } from "react";
-import { Link, NavLink } from "react-router-dom";
-
-
-const Category=(arr)=>{
-    const [categories, setCategories] = useState([]);
-
-    useEffect(()=>{
-        fetch("https://fakestoreapi.com/products/categories").then(res=> res.json())
-        .then(data => setCategories(data))
-    }, [])
-    return(
-        <>
-        {categories.map(cat=>(
-            <NavLink key={cat} to={`/product/category/${cat}`}>
-                <ul>
-                    <li>{cat}</li>
-                </ul>
-                </NavLink>
-        ))}
-        </>
-    )
-}
-
-export default Category;
