--- vite-project/src/sass/components/_class.scss
@@ -1,16 +1,0 @@
-.titles{
-    width: 100%;
-    text-align: center;
-    font-weight: 600;
-    font-size:large;
-    color: $spanColor1
-}
-
-.carrito{
-    @include flex(column, center, center);
-    .cartForm{
-        input{
-            border: 1px solid $bgPrincipal;
-        }
-    }
-}
