--- vite-project/src/sass/components/_container.scss
@@ -1,60 +1,0 @@
-.container{
-    width: auto;
-    flex-wrap: wrap;
-    @include flex(auto, center, center);
-}
-
-.item{
-    @include flex(column, center, center);
-    width: 300PX;
-    height: 450px;
-    margin: 5px;
-    border-radius: 10px;
-    border: 1px solid white;
-    background-color: $textSecundary;
-    h3{
-        text-align: center;
-        font-size: 15px;
-        padding: 5px;
-    }
-
-    img{
-        height: 250px;
-        width: auto;
-        border-radius: 15px;
-        padding: 4px;
-
-    }
-
-    ul{
-        @include flex(row, space-between, center);
-        background-color: $spanColor1;
-        width: 100%;
-        list-style: none;
-        border-radius: 0px 10px 10px 0px;
-        padding: 3px;
-        li{
-            button{
-                border: 1px solid $bgPrincipal;
-                padding: 5px;
-                border-radius: 10px;
-                transition: .4s;
-                &:active{
-                    transition: .3s;
-                    background-image: linear-gradient(#88B7B5,#847996);
-                }
-            }
-        }
-    }
-
-    p{
-        overflow: auto;
-        box-sizing: border-box;
-        padding: 2px;
-        text-align: center;
-        width: auto;
-    }
-    p::-webkit-scrollbar{
-        color: transparent;
-        }
-}
