--- vite-project/src/sass/components/_footer.scss
@@ -1,5 +1,0 @@
-.footer{
-    position:static;
-    bottom: 0;
-    width: 100%;
-}
