--- vite-project/src/sass/components/_navBar.scss
@@ -1,72 +1,0 @@
-@import "../utilities/colors";
-header{
-    display: flex;
-    text-align: center;
-    justify-content: space-between;
-    align-items: center;
-    top: 0;
-    left: 0;
-    padding: 15px 10%;
-    width: 100%;
-    background-color: $bgSecundary;
-    h1{
-        color: $textPrimary;
-        width: 25%;
-        display: flex;
-        align-items: center;
-        span{
-            color: $spanColor1;
-        }
-    }
-    .navBar{
-        width: 75%;
-        display: flex;
-        text-align: center;
-        align-items: center;
-        justify-content: space-between;
-        a{
-            text-decoration: none;
-            color: $textSecundary;
-            transition: .3s;
-            text-transform: capitalize;
-            &:hover{
-                background-color: $bgPrincipal;
-                padding: 5px;
-                border-radius: 1px 50%;
-                transition: .3s;
-            }
-        }
-        .cart{
-            display: flex;
-            align-items: center;
-            cursor: pointer;
-            transition: .3s;
-            .cartTotalProduct{
-                text-align: right;
-                margin: 0px;
-                color: $textPrimary;
-                background-color: $bgPrincipal;
-                border-radius: 50%;
-                padding: 3px;
-                transition: .5s;
-            }
-            .cartIcon{
-                height: 30px;
-            }
-            &:hover{
-                transition: .3s;
-                transform: scale(1.1);
-                .cartTotalProduct{
-                    transition: .3s;
-                    transform: translateX(-20px);
-                    margin-bottom: 8px;
-                    background-color: transparent;
-                    border-radius: 5px;
-                    padding: 0px;
-                    color: $bgPrincipal;
-                    font-weight: 600;
-                }
-            }
-        }
-    }
-}
