--- vite-project/src/sass/index.scss
@@ -1,28 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@import "./utilities/grid.scss";
-@import "./utilities/_mixin.scss";
-@import "./utilities/colors";
-
-@import "./components/navBar";
-@import "./components/container";
-@import "./components/productDetails";
-@import "./components/footer";
-@import "./components/class";
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;500;600&display=swap');
-
-*{
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    font-family: "Poppins";
-}
-body{
-    height: 100vh;
-    background-color: $bgPrincipal;
-    #root{
-        height: 100vh;
-    }
-}
