--- vite-project/src/sass/utilities/_colors.scss
@@ -1,5 +1,0 @@
-$bgPrincipal: #333D48;
-$bgSecundary: #4C5B6B;
-$textPrimary: #F5F5F5;
-$textSecundary: #E0E0E0;
-$spanColor1: #DFD27C;
