--- vite-project/src/sass/utilities/_grid.scss
@@ -1,21 +1,0 @@
-#root{
-    height: 100vh;
-    display: grid;
-    grid-template-columns: 1;
-    grid-template-rows: 1fr 4fr 1fr;
-    grid-template-areas: 
-    "header"
-    "main"
-    "footer"
-    ;
-}
-
-.header{
-    grid-area: header;
-}
-.main{
-    grid-area: main;
-}
-.footer{
-    grid-area: footer;
-}
