--- vite-project/src/sass/utilities/_mixin.scss
@@ -1,6 +1,0 @@
-@mixin flex($flex-direction, $justify-content, $align-items){
-    display: flex;
-    flex-direction: $flex-direction;
-    justify-content: $justify-content;
-    align-items: $align-items;
-}
